+++ src/tools/dev_utils/swinging_door.cc
@@ -18,7 +18,7 @@
 //   ----   ----      ----           -----------
 //   000    08-19-14  Rehak          New
 //   001    04-27-154 Halley Gotway  List and format output files
-//                                    
+//
 ////////////////////////////////////////////////////////////////////////
 
 using namespace std;
@@ -112,12 +112,12 @@ int main(int argc, char *argv[])
   grib_code_arg = atoi(cline[2]);
   message_type_arg = cline[3];
   station_id_arg = cline[4];
-   
+
   // Now, run the algorithm
 
   if (!process_file())
     return -1;
-   
+
   return 0;
 }
 
@@ -130,7 +130,7 @@ bool process_file()
   MetNcFile nc_file(ncfile_arg);
 
   vector< SDObservation > observations;
-  
+
   if (!nc_file.readFile(grib_code_arg, station_id_arg, message_type_arg,
                         observations))
     return false;
@@ -139,25 +139,25 @@ bool process_file()
 
   if (!write_observations(observations, "obs.txt"))
     return false;
-  
+
   // Run  the algorithm
 
   vector< SDObservation > compressed_observations;
-  
+
   if (!run_algorithm(observations, compressed_observations))
     return false;
-  
+
   // Write the compressed observations to the output file
 
   if (!write_observations(compressed_observations,
                           "compressed_obs.txt"))
     return false;
-  
+
   // Write the compressed observations as ramps
 
   if (!write_ramps(compressed_observations, "ramps.txt"))
     return false;
-  
+
   return true;
 }
 
@@ -170,10 +170,10 @@ bool run_algorithm(const vector< SDObservation > &observations,
   // Run the algorithm
 
   vector< pair< SDObservation, SDObservation > > ramps;
- 
+
   if (!compute_swinging_door_ramps(observations, error_arg, ramps))
     return false;
-  
+
   // Create the compressed observations from the ramps
 
   vector< pair< SDObservation, SDObservation > >::const_iterator ramp;
@@ -182,7 +182,7 @@ bool run_algorithm(const vector< SDObservation > &observations,
     compressed_observations.push_back(ramp->first);
     compressed_observations.push_back(ramp->second);
   }
-  
+
   return true;
 }
 
@@ -203,10 +203,10 @@ void usage()
 
        << "\t\t\"error_value\" is the error value (E) to use when "
        << "computing the corridors (required).\n"
-    
+
        << "\t\t\"grib_code\" is the grib code of the field to use "
        << "(required).\n"
-    
+
        << "\t\t\"message_type\" is the message type of the observations "
        << "to use (required).\n"
 
@@ -214,7 +214,7 @@ void usage()
        << "observations to use (required).\n"
 
        << flush;
-  
+
   exit(1);
 }
 
@@ -227,11 +227,11 @@ bool write_observations(const vector< SDObservation > &observations,
 
   mlog << Debug(2)
        << "Writing: " << file_path << "\n";
-  
+
   // Open the output file
 
   FILE *output_file;
-  
+
   if ((output_file = fopen(file_path.c_str(), "w")) == 0)
   {
     mlog << Error << "\n" + method_name + " -> "
@@ -242,19 +242,19 @@ bool write_observations(const vector< SDObservation > &observations,
   // Write the header line
 
   fprintf(output_file, "valid_time,unix_time,value\n");
-    
+
   // Write the observations
 
   vector< SDObservation >::const_iterator obs;
   for (obs = observations.begin(); obs != observations.end(); ++obs)
     fprintf(output_file, "%s,%d,%f\n",
             unix_to_yyyymmdd_hhmmss(obs->getValidTime()).text(),
-            obs->getValidTime(), obs->getValue());
+            (int) obs->getValidTime(), obs->getValue());
 
   // Close the output file
 
   fclose(output_file);
-  
+
   return true;
 }
 
@@ -267,22 +267,22 @@ bool write_ramps(const vector< SDObservation > &observations,
 
   mlog << Debug(2)
        << "Writing: " << file_path << "\n";
-  
+
   // Open the output file
 
   FILE *output_file;
-  
+
   if ((output_file = fopen(file_path.c_str(), "w")) == 0)
   {
     mlog << Error << "\n" + method_name + " -> "
          << "Error opening output file: " << file_path << ".\n\n";
     return false;
   }
-    
+
   // Write the header line
 
   fprintf(output_file, "start_time,end_time,run_secs,rise,slope\n");
-  
+
   // Write the observations
 
   if (observations.size() % 2 != 0)
@@ -292,28 +292,28 @@ bool write_ramps(const vector< SDObservation > &observations,
          << "An odd number were found, so there is probably a programming error. "
          << "Printing out the ramps, ignoring the last observation.\n\n";
   }
-  
+
   size_t num_ramps = observations.size() / 2;
-  
+
   for (size_t i = 0; i < num_ramps; ++i)
   {
     size_t start_obs = i * 2;
     size_t end_obs = start_obs + 1;
-    
+
     int run_secs = observations[end_obs].getValidTime() - observations[start_obs].getValidTime();
     double rise = observations[end_obs].getValue() - observations[start_obs].getValue();
     double slope = rise / (double)run_secs;
-    
+
     fprintf(output_file, "%s,%s,%d,%f,%f\n",
             observations[start_obs].getValidTimeString().c_str(),
             observations[end_obs].getValidTimeString().c_str(),
             run_secs, rise, slope);
   }
-  
+
   // Close the output file
 
   fclose(output_file);
-  
+
   return true;
 }
